Update calliope-mini-3_2_1_breadboard.svg
A member of the Calliope team found an error in the visual representation. The LEDs were not displayed, so I adjusted that. The fzpz file worked well and was tested in the forum. Now I’ve tried to make the change in the pull request. Unfortunately, my Inkscape generates different XML code. Each object has a "style" entry. Could this be a problem?